Support specifying timezone
By default the 'BST' timezone is used, which is incorrect for most
of the world. By specifying the 'timezone' variable in the conf
file, a different time zone can be selected.
This can be either an actual timezone, e.g. "America/Los_Angeles",
or the keyword 'auto', which will try to find the correct timezone
via geolocation.