Segment RecentClips by day in Samba
In the Samba share, create folders for each day of Sentry recordings
instead of putting them all directly in the RecentClips folders, which
can lead to very long load times for the Samba share.
Also clean up folders in the Samba share when they become empty
after releasing a snapshot.