Remove support for non-headless setup
For now, HEADLESS_SETUP=true still needs to be set in the configuration
file because the rc.local in the release image uses it, but manual
setup is no longer supported by the setup script.
Also rework "setup-teslausb upgrade" so that the entire upgrade process
is run from the new version.