Clean the apt cache and run fstrim on /
Clean the apt cache before package upgrades begin (so there's more
free space for the upgrade, which is especially needed for kernel
upgrades), and then again after setup is done to remove all the
cached package files that are no longer needed.
Run fstrim after setup to notify the storage device about blocks
that are no longer in use.