Set vm and cpu parameters for better performance
Set dirty_background_bytes, the amount of dirty memory at which
the kernel starts flushing to disk in the background, to 64KB.
This ensures writing to disk starts very soon after the car
begins writing to the emulated drive.

Set dirty_ratio, the amount of dirty memory allowed to accumulate
before triggering blocking foreground writes, to 80 (percent).
This prevents the file-storage process from blocking when the
car writes to the emulated drive.

Set scaling_governor to "conservative". This leaves the CPU idling
at low frequency most of the time, but allows it to be clocked
faster when the car writes to the emulated drive, which increases
throughput.