Add optional flags for rclone sync
Allow specifying additional flags for rclone sync, e.g. "--checksum" to 
compare source and target by checksum instead of size and time, which 
some targets, like Nextcloud, apparently need because they set the 
size and date of the destination file before actually transferring it, 
leading to data loss if the transfer is interrupted.