Rework snapshot maintenance, increase minimum sd size
Tesla wants the emulated drive to be at least 30GB, we need additional
space on top of that, so increase the minimum required storage medium
size to 64 GB.

Take the new snapshot before cleaning up old snapshots so that links
to the previous snapshot can be updated with links to the new snapshot
before deleting the old snapshot. This prevents files from temporarily
disappearing from the merged view in /mutable/TeslaCam/

Use 9 GB instead of the allocated size of the image as the amount of
free space to try and maintain. This should be enough to allow the car
to record until the next snapshot without wasting too much space.
(previously this was intended to use the size of the image, but
actually used the allocated size of the image, which could be much
smaller and potentially not leave enough space for another hour of
recordings)