Keep more free space
Previously the backingfiles partition could fill up completely
if recordings for the last hour averaged 38 MB each. While this
is unlikely, it's possible, and would prevent the car from doing
any further writes, even if the "CAM" drive appeared empty to the
car.
Increase the free space reserve to prevent this from happening.