Workaround for Pi presenting as unresponsive USB device
If only the dwc2 module is loaded at boot then the Pi will present
itself as a non-responsive USB device until a gadget is created.
When connected to a Windows host, this will show as a device with
vendor ID 0 (which is invalid), and Windows might pop up an error
message saying "USB device not recognized", while a connected Linux
host will log messages like "device descriptor read/64, error -110"
to the kernel log, and 'lsusb' might hang for over a minute or until
a gadget is created.
To work around this, make sure the g_ether module is loaded along
with dwc2. The archiveloop script will then unload the g_ether
module (the attached host will see no device at all at this point)
before creating the mass storage gadget.